refactor: add BaseComponentUtils
